+++ index.html
@@ -232,7 +232,7 @@
 
   <!-- Скрипты -->
   <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
