+++ js/data.js
@@ -1,0 +1,38 @@
+export const TITLES = 'Вас также может заинтересовать:';
+
+export const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+export const CHECKINS = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+export const CHECKOUTS = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+export const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+export const DESCRIPTIONS = 'Лучше места для проживания и не придумать.';
+
+export const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
